Test relaese

+++ database/migrations/2022_09_20_000006_create_invoice_payment_item_table.php
@@ -31,7 +31,7 @@ return new class extends Migration
                 ->comment('Piece of paid from total payment');
 
             $table
-                ->tinyInteger('percent')
+                ->integer('percent')
                 ->comment('Piece of percent from total payment');
 
             $table->dateTime('payment_date');
